Export internal links
